Move methods in `transaction/address.ts` to `address/address.ts`
