+++ src.ts/hash/message.ts
@@ -1,6 +1,6 @@
 import { keccak256 } from '../crypto/index.js';
 import { MessagePrefix } from '../constants/index.js';
-import { recoverAddress } from '../transaction/index.js';
+import { recoverAddress } from '../address/index.js';
 import { concat } from '../utils/index.js';
 import { toUtf8Bytes } from '../encoding/index.js';
