+++ src.ts/hash/typed-data.ts
@@ -1,7 +1,7 @@
 //import { TypedDataDomain, TypedDataField } from "@quaisproject/providerabstract-signer";
 import { getAddress } from '../address/index.js';
 import { keccak256 } from '../crypto/index.js';
-import { recoverAddress } from '../transaction/index.js';
+import { recoverAddress } from '../address/index.js';
 import {
     concat,
     defineProperties,
