+++ src.ts/wallet/hdwallet.ts
@@ -6,7 +6,7 @@
 import { computeHmac, randomBytes, SigningKey, sha256, ripemd160 } from '../crypto/index.js';
 import { Provider } from '../providers/index.js';
 import { VoidSigner } from '../signers/index.js';
-import { computeAddress } from '../transaction/index.js';
+import { computeAddress } from '../address/index.js';
 import {
     concat,
     isBytesLike,
